Add a user id to the user class.
This is necessary to avoid mapping the survey directly to the user 
object. Removing the user object will leave the survey connected to a 
pseydonymous user id instead.

The same user id is used on the data server, so that the survey results 
and the fetched data are mapped to the same id